First commit
So this here is a thing that reads events that have been shoved by the
https://github.com/ramuuns/eventcollector into a sqlite database, that
contains errors and warnings, and then shows them to you.

For now it'll just display the last two days of stuff, but like there
should be a filter form and stuff like that, but since I ain't got no
data on my local machine, for further development I need to put this on
a computer that would actually have some data, so that there's stuff to
work with.