Make frog send events about itself
which means we can now see it's own errors in itself

So the first part of all this is the events.
The way that works is that it's a `GenServer`, so basically at the start
of the thing you `start_link`, do the thing, and at the end of it you
finalize/send/cleanup (which stops that instance)

Then there's the recording itself
for regular old http requests it's rather trivial
we just override the `call` function of the endpoint, and then we wrap
the call to it's parent with a `try rescue` block which also lets us
easily catch errors and shove them in the event that we just started.
We also shove the pid of the event into the connection so that we can
access it from any other functions down the line (e.g. the bit where we
set the action in the routing pipeline once we already know what's gonna
be handling this)

The websocket stuff is more ... interesting
There's not really a way to wrap a function that would do the thing, so
we have to resort to more fun ways of doing
luckily for us the Phoenix LiveView framework does wrap some of the
processing stuff before it hits our code in their own
`:telemetry.span`s, so we can hook into the `:start` and `:stop` events
of those. Also because for whatever reason the same code is triggered
during a regular http request, we need to ignore those (we do this by
looking at the transport_pid in the socket, which is null if we're not
connected to a websocket)

Now we then create our event when we get a `:start` event from `:telemetry`
and finalize/send/cleanup the event when we get a `:stop`. We assume
that each socket (based on the id) can only have one thing running at a
time (which is a fair assumption given that it's a GenServer and it can
only handle one incomming message/data transformation at a time, so
things shouldn't interact), we then shove the event pid in a hash based
on the socket id and retreive it at the end.

however errors might happen and the result of that is that 1) we won't
get a `:stop` event, and 2) we can't just wrap the execution in a try.
Luckily for us, elixir provides a way to monitor a process and we can
find out if it got killed and why. So that is what we do: we monitor the
socket.root_pid (which is the thing that dies if you trigger an error in
your application code) when we get a message that this process has been
killed, we fetch the socked id tied to said pid (we have a map for this
in our GenServer), and thus we get the event, and thus we can shove the
error into the event, and then finalize/send/cleanup the event.

Then there's the sketchy part about monitoring the queries. So it's
pretty easy to hook into the `[:appname, :repo, :query]` events,
annoyingly they lack any useful context, so the matching of the queries
to actual events is a bit of a leap of faith. Essentially, when we
create our event, we setup a monitor for the `:query` events, and pass
it the event pid in the options. We then shove the query into that event
(and hope that it's actually coming from that event), and then before
sending the event we stop monitoring the `:query` events. I am
absolutely certain that this won't actually work properly when the
application is being hit by parallel requests that happen at the
"sameish" time and both sockets are processed in parallel. However it' s
probably good enough for the initial version.

Alternatives would be to force the user call a different function
whenever they want to query it that would wrap the query in a context
that contains the socket.id/event_pid, which would let us to definitely assign it
to the propper event, however this approach feels ugly both from an api
perspective (you now have to remember to call this other function and
pass it this argument) and from implementation perspective (you have to
write those funcitons somewhere).