Ensure we don't overflow the bucket size
seems like we might overflow the bucket when getting lots of errors,
and that leads to a place where we cannot see the errors, because our
own application is broken. not great.

So we fix that by ensuring that we never try to write to a thing
that's larger than the max index of our buckets tuple